Moved the new tests pulled from Prashant's PR to the appropriate package and started fixing them.
