--- transport/wseb/src/test/java/org/kaazing/gateway/transport/wseb/specification/BinaryIT.java
@@ -1,95 +1,0 @@
-/*
- * Copyright 2014, Kaazing Corporation. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kaazing.gateway.transport.wseb.specification;
-
-import static java.util.concurrent.TimeUnit.SECONDS;
-
-import java.net.URI;
-
-import org.junit.Ignore;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.DisableOnDebug;
-import org.junit.rules.RuleChain;
-import org.junit.rules.TestRule;
-import org.junit.rules.Timeout;
-import org.kaazing.gateway.server.test.GatewayRule;
-import org.kaazing.gateway.server.test.config.GatewayConfiguration;
-import org.kaazing.gateway.server.test.config.builder.GatewayConfigurationBuilder;
-import org.kaazing.k3po.junit.annotation.Specification;
-import org.kaazing.k3po.junit.rules.K3poRule;
-import org.kaazing.test.util.MethodExecutionTrace;
-
-public class BinaryIT {
-
-    private TestRule trace = new MethodExecutionTrace();
-    private K3poRule k3po = new K3poRule().setScriptRoot("org/kaazing/specification/wse/data/binary");
-    private final TestRule timeout = new DisableOnDebug(new Timeout(15, SECONDS));
-
-    private GatewayRule gateway = new GatewayRule() {
-        {
-         // @formatter:off
-            GatewayConfiguration configuration =
-                    new GatewayConfigurationBuilder()
-                        .service()
-                            .accept(URI.create("wse://localhost:8080/path"))
-                            .type("echo")
-                        .done()
-                    .done();
-            // @formatter:on
-            init(configuration, "log4j-trace.properties");
-        }
-    };
-
-    @Rule
-    public TestRule chain = RuleChain.outerRule(trace).around(timeout).around(k3po).around(gateway);
-
-    @Ignore("X-WebSocket-Version missing in handshake response")
-    @Test
-    @Specification("echo.payload.length.0/request")
-    public void shouldEchoFrameWithPayloadLength0() throws Exception {
-        k3po.finish();
-    }
-
-    @Ignore("X-WebSocket-Version missing in handshake response")
-    @Test
-    @Specification("echo.payload.length.127/request")
-    public void shouldEchoFrameWithPayloadLength127() throws Exception {
-        k3po.finish();
-    }
-
-    @Ignore("X-WebSocket-Version missing in handshake response")
-    @Test
-    @Specification("echo.payload.length.128/request")
-    public void shouldEchoFrameWithPayloadLength128() throws Exception {
-        k3po.finish();
-    }
-
-    @Ignore("X-WebSocket-Version missing in handshake response")
-    @Test
-    @Specification("echo.payload.length.65535/request")
-    public void shouldEchoFrameWithPayloadLength65535() throws Exception {
-        k3po.finish();
-    }
-
-    @Ignore("X-WebSocket-Version missing in handshake response")
-    @Test
-    @Specification("echo.payload.length.65536/request")
-    public void shouldEchoFrameWithPayloadLength65536() throws Exception {
-        k3po.finish();
-    }
-}
